AI_devs4/se02e02/code/pyproject.toml
[project]
name = "electricity-agent"
version = "0.1.0"
description = "Agent solving the electricity cable-rotation puzzle via OpenRouter vision"
requires-python = ">=3.11"
dependencies = [
    "httpx>=0.27",
    "pillow>=10.0",
    "python-dotenv>=1.0",
    "rich>=13.0",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src"]

[tool.uv]
dev-dependencies = [
    "ruff>=0.4",
]
